Workaround new window LTI launches.
When launching a LTI tool in a new window directly from Canvas it will not have a window.opener set because Canvas launches the new window to it's own URL (https://canvas.instructure.com...) which then redirects to the tool. Browsers block the window.opener property if it doesn't match the domain that the new window was opened with and in this case a LTI tool will be on a different domain.

This doesn't introduce a security issue because the checking of state is done in step 3 and if it was successfully set by a cookie the LTI launch will continue without a problem, but it does remove the pause that is seen when we don't have a valid window to send our post messages to.